222791, 222795 - Added key filtering to the text field, and a number formatter to aid in presentation.

+++ bundles/org.eclipse.swt/Eclipse SWT PI/cocoa/org/eclipse/swt/internal/cocoa/NSString.java
@@ -40,8 +40,8 @@ public int /*long*/ UTF8String() {
 	return OS.objc_msgSend(this.id, OS.sel_UTF8String);
 }
 
-public short characterAtIndex(int /*long*/ index) {
-	return (short)OS.objc_msgSend(this.id, OS.sel_characterAtIndex_, index);
+public int /*long*/ characterAtIndex(int /*long*/ index) {
+	return OS.objc_msgSend(this.id, OS.sel_characterAtIndex_, index);
 }
 
 public void getCharacters(char[] buffer) {
